+++ classtest2/src/main/java/com/example/classtest2/repo/CollectionRepo.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CollectionRepo extends JpaRepository<Collection,Long> {
+public interface CollectionRepo extends JpaRepository<Collection,Double> {
 
 }
